requirements.txt
accelerate==1.11.0
ag2==0.9.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alibabacloud-credentials==1.0.3
alibabacloud-credentials-api==1.0.0
alibabacloud-ecs20140526==7.2.5
alibabacloud-tea==0.4.3
alibabacloud-tea-openapi==0.4.1
alibabacloud-tea-util==0.3.14
alibabacloud_endpoint_util==0.0.4
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
apache-tvm-ffi==0.1.2
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
audioread==3.1.0
av==16.0.1
azure-core==1.36.0
azure-identity==1.25.1
azure-mgmt-compute==37.1.0
azure-mgmt-core==1.6.0
azure-mgmt-network==30.0.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blake3==1.0.8
bleach==6.3.0
blinker==1.9.0
borb==3.0.3
boto3==1.41.0
botocore==1.41.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cli_exit_tools==1.2.7
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
compressed-tensors==0.12.2
contourpy==1.3.2
cryptography==44.0.3
cssselect==1.3.0
cuda-bindings==13.0.3
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
darabonba-core==1.0.4
dashscope==1.25.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
dotenv==0.9.9
easyocr==1.7.2
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
evdev==1.9.2
executing==2.2.1
fabric==3.2.2
Farama-Notifications==0.0.4
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.4.0
fastar==0.6.0
fastdtw==0.3.4
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.20.0
flashinfer-python==0.5.2
Flask==3.0.3
fonttools==4.60.1
formulas==1.3.3
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
func_timeout==4.3.5
gdown==5.2.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.76.0
grpcio-status==1.71.2
gymnasium==0.28.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageHash==4.3.2
ImageIO==2.37.2
interegular==0.3.3
invoke==2.2.1
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
JSON-minify==0.3.0
json5==0.12.1
json_repair==0.54.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
lib_detect_testenv==3.0.0
librosa==0.11.0
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
mistune==3.1.4
model-hosting-container-standards==0.1.9
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
msgspec==0.19.0
msrest==0.7.1
multidict==6.7.0
multiprocess==0.70.18
mutagen==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
notebook==7.5.0
notebook_shim==0.2.4
numba==0.61.2
numpy>=1.24,<2.3
numpy-financial==1.0.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.2.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauth2client==4.1.3
oauthlib==3.3.1
odfpy==1.4.1
openai==2.8.1
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outlines_core==0.2.11
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
playwright==1.56.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyacoustid==1.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
PyDrive==1.3.1
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMuPDF==1.26.6
PyNaCl==1.6.1
pynput==1.7.8
pyparsing==3.2.5
pypdf==6.3.0
PyPDF2==3.0.1
pypdfium2==5.0.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-pptx==1.0.2
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
RapidFuzz==3.14.3
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-click==1.9.4
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.7.0
schedula==1.5.70
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.45.0
setproctitle==1.3.7
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.65.2
traitlets==5.14.3
transformers==4.57.3
triton==3.5.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.1
wandb==0.23.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==2.0.1
wrapt_timeout_decorator==1.5.1
xformers==0.0.33.post1
xgrammar==0.1.25
xlsxwriter==3.2.9
xmltodict==1.0.2
yarl==1.22.0
datasets>=2.14.0
